Move rules classes into rules package
