+++ src/main/java/com/ericsson/ei/handlers/DownstreamIdRulesHandler.java
@@ -18,6 +18,7 @@ package com.ericsson.ei.handlers;
 
 import java.util.List;
 
+import com.ericsson.ei.rules.MatchIdRulesHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
