+++ src/main/java/com/ericsson/ei/handlers/EventHandler.java
@@ -16,6 +16,7 @@
 */
 package com.ericsson.ei.handlers;
 
+import com.ericsson.ei.rules.IdRulesHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.amqp.core.Message;
