+++ src/main/java/com/ericsson/ei/handlers/ExtractionHandler.java
@@ -16,6 +16,7 @@
  */
 package com.ericsson.ei.handlers;
 
+import com.ericsson.ei.rules.ProcessRulesHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
