+++ src/test/java/com/ericsson/ei/waitlist/TestWaitListWorker.java
@@ -39,7 +39,7 @@ import org.springframework.amqp.rabbit.connection.CachingConnectionFactory;
 import org.springframework.amqp.rabbit.core.RabbitAdmin;
 
 import com.ericsson.ei.handlers.EventToObjectMapHandler;
-import com.ericsson.ei.handlers.MatchIdRulesHandler;
+import com.ericsson.ei.rules.MatchIdRulesHandler;
 import com.ericsson.ei.handlers.RMQHandler;
 import com.ericsson.ei.jmespath.JmesPathInterface;
 import com.ericsson.ei.mongo.MongoDBHandler;
